requirements.txt
streamlit==1.30.0
numpy==1.26.4
Pillow==10.3.0
tensorflow==2.16.1
matplotlib==3.8.4